* Improved Import screen, added "Test Import" option
git-svn-id: https://svn.eprints.org/eprints/trunk/system@2281 9491667e-5006-0410-a446-efbe8990b998

+++ perl_lib/EPrints/Plugin/Screen/Import.pm
@@ -15,57 +15,88 @@ sub new
 
 	my $self = $class->SUPER::new(%params);
 
-	$self->{actions} = [qw/ dryrun import /];
+	$self->{actions} = [qw/ test import /];
 
-	# TODO uncomment when ready for primetime
-	#$self->{appears} = [
-	#	{
-	#		place => "item_tools",
-	#		action => "import",
-	#		position => 200,
-	#	}
-	#];
+	$self->{appears} = [
+		{
+			place => "item_tools",
+			position => 200,
+		}
+	];
 
 	return $self;
 }
 
+sub properties_from
+{
+
+	my( $self ) = @_;
+	
+	$self->SUPER::properties_from;
+
+	my $pluginid = $self->{session}->param( "pluginid" );
+	
+	if( defined $pluginid )
+	{
+		my $plugin = $self->{session}->plugin( $pluginid, dataset=>$self->{session}->get_repository->get_dataset( "inbox" ) );
+		if( !defined $plugin || $plugin->broken )
+		{
+			$self->{processor}->add_message( "error", $self->{session}->html_phrase( "general:bad_param" ) );
+			return;
+		}
+
+		my $req_plugin_type = "list/eprint";
+		unless( $plugin->can_produce( $req_plugin_type ) )
+		{
+			$self->{processor}->add_message( "error", $self->{session}->html_phrase( "general:bad_param" ) );
+			return;
+		}
+
+		$self->{processor}->{plugin} = $plugin;
+
+	}
+
+}
+
+sub allow_test
+{
+	my( $self ) = @_;
+	return $self->allow( "create_eprint" );
+}
+
 sub allow_import
+{
+	my( $self ) = @_;
+	return $self->allow_test;
+}
+
+sub action_test
 {
 	my ( $self ) = @_;
 
-	return $self->allow( "create_eprint" );
+	$self->_import( 1 );
 }
 
 sub action_import
 {
-	my( $self ) = @_;
+	my ( $self ) = @_;
+
+	$self->_import( 0 );
+	$self->{processor}->{screenid} = "Items";
+}
+
+sub _import
+{
+	my( $self, $dryrun ) = @_;
 
 	my $session = $self->{session};
 	my $ds = $session->get_repository->get_dataset( "inbox" );
 
-	my $pluginid = $session->param( "pluginid" );
-	my $plugin = $session->plugin( $pluginid, dataset=>$ds ); #, parseonly=>1 );
-	if( !defined $plugin || $plugin->broken )
-	{
-		$self->{processor}->add_message( "error", $session->html_phrase( "general:bad_param" ) );
-		return;
-	}
-
-	my $req_plugin_type = "list/eprint";
-	unless( $plugin->can_produce( $req_plugin_type ) )
-	{
-		$self->{processor}->add_message( "error", $session->html_phrase( "general:bad_param" ) );
-		return;
-	}
-
-
-	my $fh = $session->{query}->upload( "importfile" );
-
-
+	# Write to temp file
+	my $fh = $self->{session}->{query}->upload( "import_filename" );
 	seek( $fh, 0, SEEK_SET );
 
 	my( $buffer );
-
 	my $tmp_file = "/tmp/eprints.import.$$";
 	open( TMP, ">$tmp_file" ) || die "Could not write to $tmp_file";
 	while( read( $fh, $buffer, 1024 ) )
@@ -74,35 +105,45 @@ sub action_import
 	}
 	close TMP;
 
+	# Build command
 	my $import_script = $EPrints::SystemSettings::conf->{base_path}."/bin/import";
 	my $ds_id = "inbox";
-	my $cmd = $import_script." --scripted ".$session->get_repository->get_id." ".$ds_id." ".$plugin->get_subtype." --user ".$self->{processor}->{user}->get_id." ".$tmp_file;
+	my $cmd = $import_script." --scripted ".$session->get_repository->get_id." ".$ds_id." ".$self->{processor}->{plugin}->get_subtype." --user ".$self->{processor}->{user}->get_id." ".$tmp_file;
+	$cmd .= " --parse-only" if $dryrun;
 
-#	print STDERR "$cmd\n";
-
+	# Run command without user check
 	my $pid = open( OUTPUT, "EPRINTS_NO_CHECK_USER=1 $cmd 2>&1|" );
 	my @imp_out = <OUTPUT>;
 	close OUTPUT;
 
-	if( -e $tmp_file )
+	# Remove temp file
+	if( !$dryrun )
 	{
-		unlink( $tmp_file );
+		if( -e $tmp_file )
+		{
+			unlink( $tmp_file );
+		}
 	}
 
 	my @misc = ();
 	my $ok = 0;
+	my $parsed = 0;
 	my @ids;
 	foreach my $line ( @imp_out )
 	{
 		if( $line !~ s/^EPRINTS_IMPORT: // )
 		{
-			push @misc,$line;
+			push @misc,$line unless $line =~ /^\s+$/s;
 			next;
 		}
 		chomp $line;
 		if( $line =~ m/ITEM_IMPORTED (\d+)/ )
 		{
 			push @ids, $1;
+		}
+		if( $line =~ m/ITEM_PARSED/ )
+		{
+			$parsed++;
 		}
 		if( $line =~ m/^DONE (\d+)$/ )
 		{
@@ -115,36 +156,52 @@ sub action_import
 		session => $session,
 		ids=>\@ids );
 
-	if( $ok && $list->count > 0)
+	if( $dryrun )
 	{
-		if( scalar @misc > 0 )
+		if( $ok )
 		{
-			my $pre = $session->make_element( "pre" );
-			$pre->appendChild( $session->make_text( join( "",$misc[0..99]) ) );
-			$self->{processor}->add_message( "warning", $pre );
+			$self->{processor}->add_message( "message", $session->html_phrase(
+				"Plugin/Screen/Import:test_completed", 
+				count => $session->make_text( $parsed ) ) );
 		}
-		$self->{processor}->add_message( "message", $session->make_text( "Imported: ".$list->count ));
-		$self->{processor}->{screenid} = "Items";
+		else
+		{
+			$self->{processor}->add_message( "warning", $session->html_phrase( 
+				"Plugin/Screen/Import:test_failed", 
+				count => $session->make_text( $parsed ) ) );
+		}
 	}
 	else
 	{
-		my $pre = $session->make_element( "pre" );
-		$pre->appendChild( $session->make_text( join( "",$misc[0..99]) ) );
-		$self->{processor}->add_message( "error", $pre );
+		if( $ok )
+		{
+			$self->{processor}->add_message( "message", $session->html_phrase( 
+				"Plugin/Screen/Import:import_completed", 
+				count => $session->make_text( $list->count ) ) );
+		}
+		else
+		{
+			$self->{processor}->add_message( "warning", $session->html_phrase( 
+				"Plugin/Screen/Import:import_failed", 
+				count => $session->make_text( $list->count ) ) );
+		}
 	}
 
-	# not used yet.
+	if( scalar @misc > 0 )
+	{
+		my $pre = $session->make_element( "pre" );
+		$pre->appendChild( $session->make_text( join( "", @misc[0..99] ) ) );
+		$self->{processor}->add_message( "warning", $session->html_phrase(
+			"Plugin/Screen/Import:import_errors",
+			errors => $pre ) );
+	}
+	
 }
 
-sub allow_dryrun
+sub redirect_to_me_url
 {
-	my ( $self ) = @_;
-
-	return $self->allow( "create_eprint" );
-} 
-
-sub action_dryrun
-{
+	my( $self ) = @_;
+	return $self->SUPER::redirect_to_me_url."&import_filename=" . $self->{session}->param( "import_filename" ) . "&pluginid=" . $self->{processor}->{plugin}->get_id;
 }
 
 sub render
@@ -159,25 +216,16 @@ sub render
 	# TODO: preamble/instructions
 
 	my $form =  $session->render_form( "post" );
+	$form->appendChild( $session->render_hidden_field( "screen", $self->{processor}->{screenid} ) );
 	$page->appendChild( $form );
 
-	# TODO: cut and paste?
-	#my $textarea = $session->make_element( "textarea", 
-	#	name => "data",
-	#	"accept-charset" => "utf-8",
-	#	wrap => "virtual",
-	#);
-	#$form->appendChild( $textarea );
-	#$form->appendChild( $session->make_element( "br" ) );
+	$form->appendChild( $session->render_upload_field( "import_filename" ) );
 
-	$form->appendChild( $session->render_upload_field( "importfile" ) );
 	$form->appendChild( $session->make_element( "br" ) );
 
-	$form->appendChild( $session->render_hidden_field( "screen", $self->{processor}->{screenid} ) );
-
 	my @plugins = $session->plugin_list( 
-				type=>"Import",
-				can_produce=>"dataobj/".$ds->confid );
+			type=>"Import",
+			can_produce=>"dataobj/".$ds->confid );
 
 	my $select = $session->make_element( "select", name => "pluginid" );
 	$form->appendChild( $select );
@@ -186,13 +234,18 @@ sub render
 	{
 		my $plugin = $session->plugin( $_ );
 		next if $plugin->broken;
-		my $opt = $session->make_element( "option", value => $_ );
+		my $opt = $session->make_element( "option", value => $_  );
+		$opt->setAttribute( "selected", "selected" ) if $self->{processor}->{plugin} && $_ eq $self->{processor}->{plugin}->get_id;
 		$opt->appendChild( $plugin->render_name );
 		$select->appendChild( $opt );
 	}
 
-	$form->appendChild( $session->render_action_buttons( import => $session->phrase( "action/import" ) ) );
+	$form->appendChild( $session->render_action_buttons( 
+		test => $session->phrase( "Plugin/Screen/Import:action:test:title" ), 
+		import => $session->phrase( "Plugin/Screen/Import:action:import:title" ) ) );
 
 	return $page;
 
 }
+
+1;
